Move _strip_extras.
This prevents a circular import since req.constructors will depend on
req.parsing.
